set time to an int64_t and fixed getting time spent

+++ headers/solver.h
@@ -1,4 +1,5 @@
 #include <vector>
+#include <iostream>
 #include <algorithm>
 
 #ifndef SOLVER_H_
@@ -8,7 +9,7 @@ typedef struct{
     int start;
     int end;
     std::vector<int> Result;
-    int time;
+    int64_t time;
 }paths;
 
 std::vector<int> Solve(std::vector<int> data, int end);
